test: “fix” test using builtin "readAnyDatabase" role
This only fixes the test insofar as the assertion now holds, but the
test throws a de.flapdoodle.reverse.TearDownException at the end.
I’m a bit perplexed as to why this works the way it does. According to
the docs at
https://www.mongodb.com/docs/manual/tutorial/create-users/#create-additional-users-for-your-deployment
we should be able to create users on other databases than "admin".
However, according to
https://www.mongodb.com/docs/manual/reference/built-in-roles/#all-database-roles
the "readAnyDatabase" (like all "all-database" roles) is only available
on the admin database. So it seems that if you want to grant a user one
of the builtin "all-database" roles, you _must_ create that user on the
"admin" database. That does strike me as a bit unintuitive to say the
least...